backend/requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI / ML
torch
torchvision
onnxruntime==1.20.1
Pillow==11.1.0
numpy==2.2.3
transformers==4.48.3
ftfy==6.3.1
regex==2024.11.6
facenet-pytorch
easyocr
opencv-python-headless

# Vector database
chromadb==0.6.3

# File metadata
exifread==3.0.0

# OCR (Phase 2, but install now)
# pytesseract==0.3.13

# Utilities
watchdog==6.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# GPU detection
gputil==1.4.0
psutil==6.1.1
